Walkthrough 01 - Course Minutes

Walkthrough 01 - Dataset on log-trace data from the number of minutes students spend on the course learning management system for online science courses

course_minutes

Format

Dataset of class tibble.

Source

John Ranellucci and Joshua Rosenberg

Examples

str(dataedu::course_minutes)

#> tibble [598 × 3] (S3: spec_tbl_df/tbl_df/tbl/data.frame)
#>  $ Bb_UserPK          : num [1:598] 44638 54346 57981 66740 67920 ...
#>  $ CourseSectionOrigID: chr [1:598] "OcnA-S116-01" "OcnA-S116-01" "OcnA-S116-01" "OcnA-S116-01" ...
#>  $ TimeSpent          : num [1:598] 1383 1191 3343 965 4095 ...
#>  - attr(*, "spec")=List of 3
#>   ..$ cols   :List of 3
#>   .. ..$ Bb_UserPK          : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   .. ..$ CourseSectionOrigID: list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_character" "collector"
#>   .. ..$ TimeSpent          : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   ..$ default: list()
#>   .. ..- attr(*, "class")= chr [1:2] "collector_guess" "collector"
#>   ..$ skip   : num 1
#>   ..- attr(*, "class")= chr "col_spec"